pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dishframed"
version = "0.1.0"
description = "Turn restaurant menu photos into a framed visual menu."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.109.0,<2",
  "pillow>=11,<12",
  "pydantic>=2.8,<3",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
]

[project.scripts]
dishframed = "dishframed.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
